Renaming schema.js to index.js.
